requirements.txt
flask
flask-cors
mediapipe
opencv-python-headless
openai
openai-whisper
gdown